[#26] Feat: MRadio 컴포넌트로 분리
